alembic v.0.7.0 is placed into requirements.txt instead of v.0.6.5
