--- alembic/versions/64103a18c68_table_photo_is_added.py
@@ -1,32 +1,0 @@
-"""table Photo is added
-
-Revision ID: 64103a18c68
-Revises: 3cfce646762
-Create Date: 2015-05-29 00:58:50.719346
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '64103a18c68'
-down_revision = '3cfce646762'
-
-from alembic import op
-import sqlalchemy as sa
-
-
-def upgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('photo',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('issue_id', sa.Integer(), nullable=True),
-    sa.Column('file_path', sa.String(length=200), nullable=True),
-    sa.ForeignKeyConstraint(['issue_id'], ['issue.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    ### end Alembic commands ###
-
-
-def downgrade():
-    ### commands auto generated by Alembic - please adjust! ###
-    pass
-    ### end Alembic commands ###
